update redirect and delete hotel

+++ src/main/java/com/example/hotel_management/Controller/HotelDetailController.java
@@ -1,13 +1,11 @@
 package com.example.hotel_management.Controller;
 
+import com.example.hotel_management.Model.BookedCapacity;
 import com.example.hotel_management.Model.DataDTO.RoomDTO;
 import com.example.hotel_management.Model.Hotel;
 import com.example.hotel_management.Model.HotelDetails;
 import com.example.hotel_management.Model.Room;
-import com.example.hotel_management.Service.HotelDetailsServices;
-import com.example.hotel_management.Service.HotelServices;
-import com.example.hotel_management.Service.RoomServices;
-import com.example.hotel_management.Service.UserServices;
+import com.example.hotel_management.Service.*;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
@@ -30,16 +28,19 @@ public class HotelDetailController {
     private final UserServices userServices;
     private final RoomServices roomServices;
     private final HotelServices hotelServices;
+    private final BookedCapacityServices bookedCapacityServices;
 
     @Autowired
     HotelDetailController(HotelDetailsServices hotelDetailsServices,
                           UserServices userServices,
                           RoomServices roomServices,
-                          HotelServices hotelServices) {
+                          HotelServices hotelServices,
+                          BookedCapacityServices BookedCapacityServices) {
         this.hotelDetailsServices = hotelDetailsServices;
         this.userServices = userServices;
         this.roomServices = roomServices;
         this.hotelServices = hotelServices;
+        this.bookedCapacityServices = BookedCapacityServices;
     }
 
     @GetMapping("/hotel-detail")
@@ -123,12 +124,12 @@ public class HotelDetailController {
             return "hotel-details";
         }
         List<Hotel> hotel = hotelServices.findByHotelID(hotel_id);
-        if (hotel.isEmpty() || hotel.get(0).getOwnerUsername() != authentication.getName()) {
-            return "redirect:/hotel-detail";
+        if (hotel.isEmpty() || !Objects.equals(hotel.get(0).getOwnerUsername(), authentication.getName())) {
+            return "redirect:/hotel-detail/your-hotel";
         }
         HotelDetails hotelDetails = hotelDetailsServices.findById(hotel_id);
 
-        model.addAttribute("hotel", hotel);
+        model.addAttribute("hotel", hotel.get(0));
         model.addAttribute("hotel_detail", hotelDetails);
 
         return "create_hotel_form";
@@ -149,24 +150,26 @@ public class HotelDetailController {
         hotelServices.saveHotel(hotel);
         hotelDetailsServices.save(hotelDetails);
 
-        return "redirect:/hotel-detail";
+        return "redirect:/hotel-detail?hotel_id=" + hotel.getHotelID();
     }
 
     @GetMapping("delete-hotel")
     public String deleteHotel(@RequestParam("hotel_id") String hotel_id, Model model) {
         List<Hotel> hotel = hotelServices.findByHotelID(hotel_id);
         if (hotel.isEmpty()){
-            return "redirect:/hotel-detail";
+            return "redirect:/hotel-detail/your-hotels";
         }
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         if (!hotel.get(0).getOwnerUsername().equals(authentication.getName()) || hotel.get(0).getIsActive() != 0) {
             return "redirect:/hotel-detail";
         }
 
-        HotelDetails hotelDetails = hotelDetailsServices.findById(hotel_id);
-        hotelDetailsServices.delete(hotelDetails.getHotelID());
-        hotelServices.deleteByHotelId(hotel_id);
-        return "redirect:/hotel-detail";
+        if (hotel.get(0).getIsActive() == 0) {
+            HotelDetails hotelDetails = hotelDetailsServices.findById(hotel_id);
+            hotelDetailsServices.delete(hotelDetails.getHotelID());
+            hotelServices.deleteByHotelId(hotel_id);
+        }
+        return "redirect:/hotel-detail/your-hotels";
     }
 
     @GetMapping("/hotel-detail/your-hotels")
@@ -210,11 +213,11 @@ public class HotelDetailController {
 
         List<Hotel> hotel = hotelServices.findByHotelID(hotel_id);
         if (hotel.isEmpty()){
-            return "redirect:/hotel-detail";
+            return "redirect:/hotel-requests";
         }
         hotel.get(0).setIsActive(1);
         hotelServices.saveHotel(hotel.get(0));
-        return "redirect:/hotel-detail";
+        return "redirect:/hotel-requests";
     }
 
     @GetMapping("/reject-hotel")
@@ -232,11 +235,11 @@ public class HotelDetailController {
 
         List<Hotel> hotel = hotelServices.findByHotelID(hotel_id);
         if (hotel.isEmpty()){
-            return "redirect:/hotel-detail";
+            return "redirect:/hotel-requests";
         }
         hotel.get(0).setIsActive(-1);
         hotelServices.saveHotel(hotel.get(0));
-        return "redirect:/hotel-detail";
+        return "redirect:/hotel-requests";
     }
 
     @GetMapping("hotel-requests")
